requirements.txt
requests
argparse
forex-python==1.8
googlesearch-python